Added classes to represent IRIT object
The IritObject.h introduces the IritWorld, IritObject, IritPolygon and
Irit point to represent the whole view, a signle object, a polygon and
a 3D point respectively.

Also implemented scanning an IritObject into these classes. Namely into
the 'world' variable which represent our view.

The API for these classes can be summerized as:
	Object = world.createObject();
	Polygon = Object.createPolygon();
	Point = Polygon.addPoint(x,y,z);