Add perspective view and change transformation pipeline
The matrix pipeline was changed to have 3 matrix multiplication
instead of 7 (or so). The 3 matrix model matches the pipeline which
was studies in class.

Also added a view matrix which translates the world into camera (view)
matrix. The matrix is set at (0, 0, -20) by default and faces (0, 0, 1)
direction.

Moreover, a perspective view was added. Two different methods of
perspective view were implemented, and the one that was studied during
lectures is used (the other uses frustum view to create perspective).
This addition introduces performance penalty as we can no longer combine
the object->projection matrix with the projection->screen matrix (since
we need to divide by 'w' after applying perspective view) and so each
point is now multiplies twice by a matrix.

And as last minor change, the wireframe color was changed to gray since I
find it easier to detect the perspective in 'perspective view'.