Added IritFigure class and tweaked mouse aided transformation
The class IritFigure was added as an intermidiate level between the
classes IritWorld and IritObject. This is because, at first, I thought
that an object represents the whole figure, but I was mistaken.
Each object is constructed out of several objects, so we need a class
representation for the whole shpill.

The addition of the IritFigure class allows us to make figure-specific
operations, like rotating, translating and scaling. This patch inroduces
a mechanism in which upon mouse key press, the coordinate of the mouse are
checked again the coordinates of the bounding boxes of all figures. If the
mouse "contained" in the bounding of one of the figures, then all transformations
are later applied to this figure.

The matrix coord_mat was added to reflect the Y ( new_y = -original_y) of
the whole world. This is because the Y axis in screen space grows down (which
is opposite to how the figure draws itself).
After changing the drawing system to bitmap, this matrix is not needed when
drawing an object, as the Y axis in the bitmap system graws up (the same as
the object's). Nevertheless, this matrix was left in the code since it helps
determine whehter the pressed key was inside a figure's bounding box.